pyproject.toml
[tool.poetry]
name = "spotify-automator"
version = "0.1.0"
description = ""
authors = ["Christopher Hoelter <git@christopherhoelter.com>"]

[tool.poetry.dependencies]
python = "^3.8"
spotipy = "^2.20.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
